+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning 2.0](https://semver.org/spec/v2
 
 ## v2.15.0-pre.2
 ### Added
+- Added new package - [AI Toolkit](https://github.com/ostris/ai-toolkit/)
 - Added Manual Install button for installing Package extensions that aren't in the indexes
 - Added Next and Previous buttons to the Civitai details page to navigate between results
 - Added Negative Rejection Steering (NRS) by @reithan to Inference
@@ -16,6 +17,9 @@ and this project adheres to [Semantic Versioning 2.0](https://semver.org/spec/v2
 ### Fixed
 - Fixed Inference custom step (e.g. HiresFix) Samplers potentially sharing state with other card UIs like model browser.
 - Fixed extension manager failing to install extensions due to incorrect clone directory
+- Fixed duplicate Python versions appearing in the Advanced Options when installing a package
+- Fixed [#1360](https://github.com/LykosAI/StabilityMatrix/issues/1360) - A1111 install not using correct torch for 5000-series GPUs
+- Fixed [#1361](https://github.com/LykosAI/StabilityMatrix/issues/1361) - numpy and other Forge startup errors
 
 ## v2.15.0-pre.1
 ### Added